Add api tester
