+++ api/handlers_test.go
@@ -5,19 +5,14 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"net/http"
-	"net/http/httptest"
 	url "net/url"
 	"strings"
 	"testing"
 	"time"
 
 	"github.com/SkynetLabs/blocker/database"
-	"github.com/SkynetLabs/blocker/skyd"
-	"github.com/sirupsen/logrus"
 	"gitlab.com/SkynetLabs/skyd/skymodules"
-	"go.mongodb.org/mongo-driver/mongo/options"
 )
 
 var (
@@ -237,9 +232,10 @@ func testHandleBlocklistGET(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
+	apiTester := newAPITester(api)
 
 	// fetch the blocklist and assert it is empty
-	bl, err := fetchBlocklist(api, nil, nil, nil)
+	bl, err := apiTester.blocklistGET(api, nil, nil, nil)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -270,7 +266,7 @@ func testHandleBlocklistGET(t *testing.T) {
 	}
 
 	// assert base case
-	bl, err = fetchBlocklist(api, nil, nil, nil)
+	bl, err = apiTester.blocklistGET(api, nil, nil, nil)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -289,7 +285,7 @@ func testHandleBlocklistGET(t *testing.T) {
 
 	// assert limit of 1
 	limit := 1
-	bl, err = fetchBlocklist(api, nil, nil, &limit)
+	bl, err = apiTester.blocklistGET(api, nil, nil, &limit)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -302,7 +298,7 @@ func testHandleBlocklistGET(t *testing.T) {
 
 	// assert offset of 1
 	offset := 1
-	bl, err = fetchBlocklist(api, nil, &offset, nil)
+	bl, err = apiTester.blocklistGET(api, nil, &offset, nil)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -318,7 +314,7 @@ func testHandleBlocklistGET(t *testing.T) {
 
 	// assert sort
 	sort := "desc"
-	bl, err = fetchBlocklist(api, &sort, nil, nil)
+	bl, err = apiTester.blocklistGET(api, &sort, nil, nil)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -341,7 +337,7 @@ func testHandleBlocklistGET(t *testing.T) {
 	hasmore := true
 	var entries []BlockedHash
 	for hasmore {
-		bl, err = fetchBlocklist(api, nil, &offset, &limit)
+		bl, err = apiTester.blocklistGET(api, nil, &offset, &limit)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -433,71 +429,4 @@ func TestVerifySkappReport(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
-}
-
-// newTestAPI returns a new API instance
-func newTestAPI(dbName string, skyd skyd.API) (*API, error) {
-	// create a nil logger
-	logger := logrus.New()
-	logger.Out = ioutil.Discard
-
-	// create database
-	db, err := database.NewCustomDB(context.Background(), "mongodb://localhost:37017", dbName, options.Credential{
-		Username: "admin",
-		Password: "aO4tV5tC1oU3oQ7u",
-	}, logger)
-	if err != nil {
-		return nil, err
-	}
-
-	// create a context with timeout
-	ctx, cancel := context.WithTimeout(context.Background(), database.MongoDefaultTimeout)
-	defer cancel()
-
-	// purge the database
-	err = db.Purge(ctx)
-	if err != nil {
-		panic(err)
-	}
-
-	// create the API
-	api, err := New(skyd, db, logger)
-	if err != nil {
-		return nil, err
-	}
-	return api, nil
-}
-
-// fetchBlocklist is a helper function that fetches the blocklist
-func fetchBlocklist(api *API, sort *string, offset, limit *int) (BlocklistGET, error) {
-	// set url values
-	values := url.Values{}
-	if offset != nil {
-		values.Set("offset", fmt.Sprint(*offset))
-	}
-	if limit != nil {
-		values.Set("limit", fmt.Sprint(*limit))
-	}
-	if sort != nil {
-		values.Set("sort", *sort)
-	}
-
-	// create the request
-	url := fmt.Sprintf("/blocklist?%s", values.Encode())
-	req := httptest.NewRequest(http.MethodGet, url, nil)
-
-	// create a recorder and execute the request
-	w := httptest.NewRecorder()
-	api.blocklistGET(w, req, nil)
-	res := w.Result()
-	defer res.Body.Close()
-
-	// handle the response
-	data, err := ioutil.ReadAll(res.Body)
-	if err != nil {
-		return BlocklistGET{}, err
-	}
-	var blg BlocklistGET
-	json.Unmarshal(data, &blg)
-	return blg, nil
 }
